requirements.txt
streamlit
tensorflow
librosa
matplotlib
pillow
numpy
soundfile
imageio-ffmpeg